Implemented a power management system

+++ Board-Application/src/adc/adc_demo.h
@@ -36,5 +36,5 @@ typedef struct _adc_result {
 */
 void Initialize_ADC_Demo(void);
 
-uint32_t home_automation_adc(uint32_t Parameter_Count, QCLI_Parameter_t *Parameter_List);
+uint16_t home_automation_adc(uint32_t Parameter_Count, QCLI_Parameter_t *Parameter_List);
